Create README.md
**Protective Web Test Project**

This project is a demo project for Protective.  The project should fulfill the provided requirements.

**Project Setup**

To setup the project you will need to create a database and the needed tables.  The project contains a scripts solution folder.  The folder contains the Protective_Test_1.0.0.sql file.  This file will generate the ProtectiveTest database and the needed tables.  It will also create the needed user roles and administrative account.

**Project Dependencies**

The following packages have been used in this project
--------------

- bootstrap 3.0.0
- Dapper 1.23
- OWIN 1.0.0
- NLog 2.1.0
- NUnit 2.6.3
- SturctureMapMVC 4.3.0.2.115

The following JQuery plugins have been used in this project
--------------

- bootstrap-toastr
- jquery datatables
- jquery form